Moved mapper to static Utilities.cs
