+++ src/Dfe.ContentSupport.Web/Models/Mapped/CsPage.cs
@@ -1,4 +1,4 @@
-﻿using Dfe.ContentSupport.Web.Services;
+﻿using Dfe.ContentSupport.Web.Common;
 using Dfe.ContentSupport.Web.ViewModels;
 
 namespace Dfe.ContentSupport.Web.Models.Mapped;
@@ -6,7 +6,7 @@ namespace Dfe.ContentSupport.Web.Models.Mapped;
 public class CsPage(ContentSupportPage contentfulPage)
 {
     public readonly List<CsContentItem> Content =
-        ContentSupportMapperService.MapContent(contentfulPage.Content);
+        Utilities.MapEntriesToContent(contentfulPage.Content);
 
     public readonly Heading Heading = contentfulPage.Heading;
     public readonly bool IsSitemap = contentfulPage.IsSitemap;
