+++ src/Dfe.ContentSupport.Web/Views/Shared/_UnsupportedElement.cshtml
@@ -1,0 +1,10 @@
+﻿@model RichTextContentItem
+
+<div style="width:100%;background-color:coral">
+    <p>
+        Unsupported element 
+        <br/>- internalName: @Model.InternalName
+        <br/>- nodeType:     @Model.NodeType
+        <br/>- value:        @Model.Value
+    </p>
+</div>
